pyproject.toml
[project]
name = "angeleyes"
version = "0.1.0"
description = "Vision-based focus and posture monitoring assistant"
requires-python = ">=3.13"
dependencies = [
    "pydantic",
    "pydantic-ai>=0.0.14",
    "click>=8.0.0",
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.0",
    "httpx>=0.25.0",
    "jinja2>=3.1.0",
    "toml>=0.10.0",
    "loguru>=0.7.0",
    "pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "pyyaml>=6.0.0",
    "pytest-asyncio>=1.0.0",
]

[project.scripts]
angeleyes = "angeleyes.cli:main"

[tool.setuptools.packages.find]
include = ["angeleyes*"]
exclude = ["llms*"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.11.6",
    "ty>=0.0.1a14",
]


[tool.ruff]
exclude = [".venv", "**/*_test.py"]

[tool.ruff.lint]
extend-select = [
  "A",      # flake8-builtins - prevent shadowing of built-in names
  "ANN",    # flake8-annotations - enforce type annotations
  "ARG",    # flake8-unused-arguments - detect unused function arguments
  "B",      # flake8-bugbear - find likely bugs and design problems
  "COM",    # flake8-commas - enforce trailing commas
  "C4",     # flake8-comprehensions - improve list/dict/set comprehensions
  "C90",    # mccabe - cyclomatic complexity checks
  "DTZ",    # flake8-datetimez - prevent naive datetime usage
  "E",      # pycodestyle errors - PEP 8 error violations
  "EM",     # flake8-errmsg - enforce error message best practices
  "F",      # pyflakes - detect various Python errors
  "FBT",    # flake8-boolean-trap - prevent boolean trap anti-pattern
  "FURB",   # refurb - modernize Python code
  "I",      # isort - import sorting and organization
  "ISC",    # flake8-implicit-str-concat - prevent implicit string concatenation
  "LOG",    # flake8-logging - logging best practices
  "G",      # flake8-logging-format - logging format string issues
  "N",      # pep8-naming - enforce PEP 8 naming conventions
  "NPY",    # numpy-specific rules - NumPy best practices
  "PD",     # pandas-vet - pandas best practices
  "PERF",   # perflint - performance anti-patterns
  "PIE",    # flake8-pie - miscellaneous improvements
  "PL",     # pylint - comprehensive Python linting
  "PT",     # flake8-pytest-style - pytest best practices
  "PTH",    # flake8-use-pathlib - prefer pathlib over os.path
  "Q",      # flake8-quotes - enforce quote style consistency
  "RET",    # flake8-return - improve return statement usage
  "RUF",    # ruff-specific rules - ruff's own additional checks
  "SIM",    # flake8-simplify - simplify code constructs
  "SLF",    # flake8-self - private member access checks
  "SLOT",   # flake8-slots - enforce __slots__ usage
  "TID",    # flake8-tidy-imports - tidy import organization
  "TC",     # flake8-type-checking - optimize type checking imports
  "UP",     # pyupgrade - upgrade syntax for newer Python versions
  "W",      # pycodestyle warnings - PEP 8 warning violations
]
extend-ignore = [
    "COM812",   # flake8-commas - allow trailing commas in function calls
    "EM102",    # flake8-errmsg - allow exception message in f-string
    "E501",     # pycodestyle - ignore line too long (let formatter handle it)
    "E501",     # pycodestyle - ignore line too long (duplicate entry)
    "PLR0913",  # pylint - allow more than 5 function arguments
    "TRY003",   # tryceratops - allow long exception messages
    "W293",     # pycodestyle - allow blank line with whitespace
]

[tool.ruff.lint.mccabe]
max-complexity = 5

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-vv --tb=short"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["angeleyes"]
omit = ["*_test.py", "*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
]

